Inicial commit for Decorator Pattern

+++ webshop/src/main/java/com/mariana/checkout/CreditCardStrategy.java
@@ -3,7 +3,7 @@ package com.mariana.checkout;
 public class CreditCardStrategy implements PaymentStrategy
 {
     @Override
-    public void pay(int amount) {
+    public void pay(Double amount) {
         String msg = String.format("Client paid %s euros with CREDIT CARD", amount);
         System.out.println(msg);
     }
