+++ webshop/src/main/java/com/mariana/checkout/PaymentStrategy.java
@@ -1,5 +1,5 @@
 package com.mariana.checkout;
 
 public interface PaymentStrategy {
-    public void pay(int amount);
+    public void pay(Double amount);
 }
